// lib/generated/consumer_responses.ts
export type RequestConversationResponse = {
  /**
   * Defines that the message is a response
   */
  kind: string;
  /**
   * The ID of the request that originated this response
   */
  reqId: string;
  /**
   * Contains one of the APIs response type names
   */
  type: string;
  /**
   * Response code according to the HTTP status codes standard
   */
  code: number;
  body: any;
  [k: string]: any;
} & {
    type?: string;
    body?: {
      /**
       * The ID of the generated conversation
       */
      conversationId: string;
      [k: string]: any;
    };
    [k: string]: any;
  };